+++ samples/lib/widgets/header.dart
@@ -6,26 +6,73 @@ import 'package:forui/forui.dart';
 import 'package:forui_samples/sample_scaffold.dart';
 
 @RoutePage()
-class HeaderPage extends SampleScaffold {
-  HeaderPage({
+class RootHeaderPage extends SampleScaffold {
+  RootHeaderPage({
     @queryParam super.theme,
   });
 
   @override
-  Widget child(BuildContext context) => Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 20),
-        child: FHeader(
-          title: 'Edit Alarm',
-          actions: [
-            FHeaderAction(
-              icon: FAssets.icons.alarmClock,
-              onPress: () {},
-            ),
-            FHeaderAction(
-              icon: FAssets.icons.plus,
-              onPress: () {},
-            ),
-          ],
-        ),
+  Widget child(BuildContext context) => FHeader(
+        title: const Text('Edit Alarm'),
+        actions: [
+          FHeaderAction(
+            icon: FAssets.icons.alarmClock,
+            onPress: () {},
+          ),
+          FHeaderAction(
+            icon: FAssets.icons.plus,
+            onPress: () {},
+          ),
+        ],
+      );
+}
+
+@RoutePage()
+class NestedHeaderPage extends SampleScaffold {
+  NestedHeaderPage({
+    @queryParam super.theme,
+  });
+
+  @override
+  Widget child(BuildContext context) => FHeader.nested(
+        title: const Text('Appointment'),
+        leftActions: [
+          FHeaderAction.back(onPress: () {}),
+        ],
+        rightActions: [
+          FHeaderAction(
+            icon: FAssets.icons.info,
+            onPress: () {},
+          ),
+          FHeaderAction(
+            icon: FAssets.icons.plus,
+            onPress: () {},
+          ),
+        ],
+      );
+}
+
+@RoutePage()
+class XNestedHeaderPage extends SampleScaffold {
+  XNestedHeaderPage({
+    @queryParam super.theme,
+  });
+
+  @override
+  Widget child(BuildContext context) => FHeader.nested(
+        title: const Text('Climate'),
+        leftActions: [
+          FHeaderAction(
+            icon: FAssets.icons.thermometer,
+            onPress: () {},
+          ),
+          FHeaderAction(
+            icon: FAssets.icons.wind,
+            onPress: null,
+          ),
+        ],
+        rightActions: [
+          FHeaderAction.x(onPress: () {}),
+        ],
       );
 }