+++ samples/lib/widgets/text_field.dart
@@ -22,7 +22,7 @@ class TextFieldPage extends SampleScaffold {
             padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 30),
             child: FTextField(
               enabled: enabled,
-              label: 'Email',
+              label: const Text('Email'),
               hint: 'john@doe.com',
             ),
           ),
@@ -44,7 +44,6 @@ class PasswordTextFieldPage extends SampleScaffold {
             padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 30),
             child: FTextField.password(
               controller: TextEditingController(text: 'My password'),
-              label: 'Password',
             ),
           ),
         ],
@@ -64,9 +63,64 @@ class MultilineTextFieldPage extends SampleScaffold {
           Padding(
             padding: EdgeInsets.symmetric(horizontal: 20, vertical: 30),
             child: FTextField.multiline(
-              label: 'Leave a review',
+              label: Text('Leave a review'),
+              maxLines: 4,
             ),
           ),
         ],
+      );
+}
+
+@RoutePage()
+class FormTextFieldPage extends SampleScaffold {
+  FormTextFieldPage({
+    @queryParam super.theme,
+  });
+
+  @override
+  Widget child(BuildContext context) => const Padding(
+        padding: EdgeInsets.all(15.0),
+        child: LoginForm(),
+      );
+}
+
+class LoginForm extends StatefulWidget {
+  const LoginForm({super.key});
+
+  @override
+  State<LoginForm> createState() => _LoginFormState();
+}
+
+class _LoginFormState extends State<LoginForm> {
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
+
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) => Form(
+        key: _formKey,
+        child: Column(
+          children: [
+            FTextField.email(
+              hint: 'janedoe@foruslabs.com',
+              help: const Text(''),
+              validator: (value) => (value?.contains('@') ?? false) ? null : 'Please enter a valid email.',
+            ),
+            const SizedBox(height: 4),
+            FTextField.password(
+              hint: '',
+              help: const Text(''),
+              validator: (value) => 8 <= (value?.length ?? 0) ? null : 'Password must be at least 8 characters long.',
+            ),
+            const SizedBox(height: 30),
+            FButton(
+              rawLabel: const Text('Login'),
+              onPress: () => _formKey.currentState!.validate(),
+            ),
+          ],
+        ),
       );
 }