+++ forui/CHANGELOG.md
@@ -1,3 +1,15 @@
+## Next
+
+### Changes
+* Add `Header.nested` widget.
+* Change `FTextField` to be usable in `Form`s. 
+* Change `FTextFieldStyle`'s default vertical content padding from `5` to `15`.
+* **Breaking** Move `FHeaderStyle` to `FHeaderStyles.rootStyle`.
+* **Breaking** Move `FHeaderActionStyle.padding` to `FRootHeaderStyle.actionSpacing`.
+* Split exports in `forui.dart` into sub-libraries.
+* **Breaking** Suffix style parameters with `Style`, i.e. `FRootHeaderStyle.action` has been renamed to `FRootHeaderStyle.actionStyle`.
+* Fix missing `key` parameter in `FTextField` constructors.
+
 ## 0.1.0
 
 * Initial release! 🚀